+++ CHANGELOG.md
@@ -27,6 +27,7 @@ All notable changes to the Docker Language Server will be documented in this fil
 - Compose
   - textDocument/completion
     - fix panic in code completion in an empty file ([#196](https://github.com/docker/docker-language-server/issues/196))
+    - fix line number assumption issues when using code completion for build targets ([#210](https://github.com/docker/docker-language-server/issues/210))
   - textDocument/hover
     - ensure results are returned even if the file has CRLFs ([#205](https://github.com/docker/docker-language-server/issues/205))
 - Bake